backend/auth/requirements.txt
# AWS SDK for Python
boto3>=1.34.0

# JWT handling
PyJWT>=2.8.0

# Cryptography for token validation and encryption
cryptography>=41.0.0

# HTTP requests for OAuth flows
requests>=2.31.0

# Property-based testing
hypothesis>=6.92.0

# Additional dependencies for production
python-dateutil>=2.8.2
